Move integration tests to test module
