+++ instrumentation/net/http/otelhttp/test/go.mod
@@ -1,0 +1,14 @@
+module go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp/test
+
+go 1.15
+
+require (
+	github.com/stretchr/testify v1.7.0
+	go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp v0.22.0
+	go.opentelemetry.io/otel v1.0.0-RC2
+	go.opentelemetry.io/otel/metric v0.22.0
+	go.opentelemetry.io/otel/oteltest v1.0.0-RC2
+	go.opentelemetry.io/otel/trace v1.0.0-RC2
+)
+
+replace go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp v0.22.0 => ../
